+++ src/hlib/libcpp/random.hpp
@@ -1,1 +1,2 @@
+#include <cstdlib>
 #include <ctime>