+++ docs/src/CHANGELOG.md
@@ -23,6 +23,7 @@
 - fix assigning `NULL` to arrays and pointers bug, #36.
 - fix `check_g++` config option bugs
 - fix not defined consts when importing packages
+- fix `random` library bug
 
 </details>